backend/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.10
pytest
bcrypt
flask_jwt_extended
requests